stroke-app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
numpy==2.1.2
Pillow==10.4.0
reportlab==4.2.5
python-multipart==0.0.9
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.7.2
xgboost==2.1.1
torch
torchvision